The KomTime class had an off-by-one bug in its handling of conversion between Date and KOM Time.
In the Test2 client, a new Command framework is introduced, to make it easier
to separate code and add new commands (in the future, it should be pretty easy
to make commands "pluggable" so you can add client commands without recompiling
the Test2 class). For now, only a few commands has been moved to the new framework,
and they can be found in the ConfCommands class. I'm not entirely sure about the
API here though, so it might (and probably will) change over time.